+++ native/tests/seal/relinkeys.cpp
@@ -39,11 +39,11 @@ namespace sealtest
                 {
                     ASSERT_EQ(keys.key(j + 2)[i].data().size(), test_keys.key(j + 2)[i].data().size());
                     ASSERT_EQ(
-                        keys.key(j + 2)[i].data().int_array().size(),
-                        test_keys.key(j + 2)[i].data().int_array().size());
+                        keys.key(j + 2)[i].data().dyn_array().size(),
+                        test_keys.key(j + 2)[i].data().dyn_array().size());
                     ASSERT_TRUE(is_equal_uint(
                         keys.key(j + 2)[i].data().data(), test_keys.key(j + 2)[i].data().data(),
-                        keys.key(j + 2)[i].data().int_array().size()));
+                        keys.key(j + 2)[i].data().dyn_array().size()));
                 }
             }
         }
@@ -69,11 +69,11 @@ namespace sealtest
                 {
                     ASSERT_EQ(keys.key(j + 2)[i].data().size(), test_keys.key(j + 2)[i].data().size());
                     ASSERT_EQ(
-                        keys.key(j + 2)[i].data().int_array().size(),
-                        test_keys.key(j + 2)[i].data().int_array().size());
+                        keys.key(j + 2)[i].data().dyn_array().size(),
+                        test_keys.key(j + 2)[i].data().dyn_array().size());
                     ASSERT_TRUE(is_equal_uint(
                         keys.key(j + 2)[i].data().data(), test_keys.key(j + 2)[i].data().data(),
-                        keys.key(j + 2)[i].data().int_array().size()));
+                        keys.key(j + 2)[i].data().dyn_array().size()));
                 }
             }
         }
@@ -94,7 +94,7 @@ namespace sealtest
                     size_t coeff_modulus_size = coeff_modulus.size();
                     size_t rns_poly_uint64_count = util::mul_safe(coeff_count, coeff_modulus_size);
 
-                    IntArray<Ciphertext::ct_coeff_type> error;
+                    DynArray<Ciphertext::ct_coeff_type> error;
                     error.resize(rns_poly_uint64_count);
                     auto destination = error.begin();